feat(Astro): Add common components and update tsconfig paths for improved module resolution
