+++ examples/astro/src/components/common/Error.astro
