+++ examples/astro/src/components/common/Header.tsx
