+++ examples/astro/src/components/common/Layout.astro
