+++ examples/astro/src/components/content-types/dotComponents.ts
@@ -1,0 +1,32 @@
+import { CustomNoComponent } from "./Empty";
+import Activity from "./Activity";
+import Banner from "./Banner";
+import BannerCarousel from "./BannerCarousel";
+import CalendarEvent from "./CalendarEvent";
+import CallToAction from "./CallToAction";
+import CategoryFilter from "./CategoryFilter";
+import DotCMSImage from "./Image";
+import PageForm from "./PageForm";
+import Product from "./Product";
+import SimpleWidget from "./SimpleWidget";
+import StoreProductList from "./StoreProductList";
+import VtlInclude from "./VtlInclude";
+import WebPageContent from "./WebPageContent";
+
+export const dotComponents = {
+    Activity,
+    Banner,
+    BannerCarousel,
+    CallToAction,
+    CategoryFilter,
+    CustomNoComponent,
+    PageForm,
+    Product,
+    SimpleWidget,
+    StoreProductList,
+    VtlInclude,
+    calendarEvent: CalendarEvent,
+    Image: DotCMSImage,
+    webPageContent: WebPageContent,
+  };
+
