+++ examples/astro/src/components/content-types/index.ts
@@ -1,0 +1,15 @@
+export * from "./Activity";
+export * from "./Banner";
+export * from "./BannerCarousel";
+export * from "./CalendarEvent";
+export * from "./CallToAction";
+export * from "./CategoryFilter";
+export * from "./dotComponents";
+export * from "./Empty";
+export * from "./Image";
+export * from "./PageForm";
+export * from "./Product";
+export * from "./SimpleWidget";
+export * from "./StoreProductList";
+export * from "./VtlInclude";
+export * from "./WebPageContent";
