--- examples/astro/src/components/content-types/index.tsx
@@ -1,31 +1,0 @@
-import { CustomNoComponent } from "./Empty";
-import Activity from "./Activity";
-import Banner from "./Banner";
-import BannerCarousel from "./BannerCarousel";
-import CalendarEvent from "./CalendarEvent";
-import CallToAction from "./CallToAction";
-import CategoryFilter from "./CategoryFilter";
-import SimpleWidget from "./SimpleWidget";
-import ImageComponent from "./Image";
-import PageForm from "./PageForm";
-import Product from "./Product";
-import StoreProductList from "./StoreProductList";
-import VtlInclude from "./VtlInclude";
-import WebPageContent from "./WebPageContent";
-
-export const pageComponents = {
-  Activity: Activity,
-  Banner: Banner,
-  BannerCarousel: BannerCarousel,
-  calendarEvent: CalendarEvent,
-  CallToAction: CallToAction,
-  CategoryFilter: CategoryFilter,
-  CustomNoComponent: CustomNoComponent,
-  Image: ImageComponent,
-  PageForm: PageForm,
-  Product: Product,
-  SimpleWidget: SimpleWidget,
-  StoreProductList: StoreProductList,
-  VtlInclude: VtlInclude,
-  webPageContent: WebPageContent,
-};
