+++ examples/astro/src/components/ui/index.ts
@@ -1,0 +1,5 @@
+export * from "./BlogCard";
+export * from "./ContactForm";
+export * from "./DestinationListing";
+export * from "./NotFound";
+export * from "./RecommendedCard";
