+++ examples/astro/src/hooks/index.ts
@@ -1,0 +1,2 @@
+export * from "./useIsEditMode";
+export * from "./useDebounce";
