+++ examples/astro/src/hooks/useIsEditMode.ts
