+++ examples/astro/src/integration/dotcms/index.ts
@@ -1,0 +1,2 @@
+export * from "./getPage";
+export * from "./dotCMSClient";
